# Speedup tests command_line integration