--- a/tests/components/command_line/test_binary_sensor.py
+++ b/tests/components/command_line/test_binary_sensor.py
@@ -206,16 +206,19 @@ async def test_updating_to_often(
     hass: HomeAssistant, caplog: pytest.LogCaptureFixture
 ) -> None:
     """Test handling updating when command already running."""
+
+    wait_till_event = asyncio.Event()
+    wait_till_event.set()
     called = []
 
     class MockCommandBinarySensor(CommandBinarySensor):
-        """Mock entity that updates slow."""
+        """Mock entity that updates."""
 
         async def _async_update(self) -> None:
-            """Update slow."""
+            """Update the entity."""
             called.append(1)
-            # Add waiting time
-            await asyncio.sleep(1)
+            # Wait till event is set
+            await wait_till_event.wait()
 
     with patch(
         "homeassistant.components.command_line.binary_sensor.CommandBinarySensor",
@@ -232,7 +235,7 @@ async def test_updating_to_often(
                             "command": "echo 1",
                             "payload_on": "1",
                             "payload_off": "0",
-                            "scan_interval": 0.1,
+                            "scan_interval": 10,
                         }
                     }
                 ]
@@ -241,23 +244,25 @@ async def test_updating_to_often(
         await hass.async_block_till_done()
 
     assert called
+    async_fire_time_changed(hass, dt_util.now() + timedelta(seconds=15))
+    wait_till_event.set()
+    asyncio.wait(0)
     assert (
         "Updating Command Line Binary Sensor Test took longer than the scheduled update interval"
         not in caplog.text
     )
-    called.clear()
-    caplog.clear()
 
-    async_fire_time_changed(hass, dt_util.now() + timedelta(seconds=1))
-    await hass.async_block_till_done()
+    # Simulate update takes too long
+    wait_till_event.clear()
+    async_fire_time_changed(hass, dt_util.now() + timedelta(seconds=10))
+    await asyncio.sleep(0)
+    async_fire_time_changed(hass, dt_util.now() + timedelta(seconds=10))
+    wait_till_event.set()
 
-    assert called
     assert (
         "Updating Command Line Binary Sensor Test took longer than the scheduled update interval"
         in caplog.text
     )
-
-    await asyncio.sleep(0.2)
 
 
 async def test_updating_manually(