--- a/tests/components/command_line/test_switch.py
+++ b/tests/components/command_line/test_switch.py
@@ -650,16 +650,19 @@ async def test_updating_to_often(
     hass: HomeAssistant, caplog: pytest.LogCaptureFixture
 ) -> None:
     """Test handling updating when command already running."""
+
     called = []
+    wait_till_event = asyncio.Event()
+    wait_till_event.set()
 
     class MockCommandSwitch(CommandSwitch):
-        """Mock entity that updates slow."""
+        """Mock entity that updates."""
 
         async def _async_update(self) -> None:
-            """Update slow."""
+            """Update entity."""
             called.append(1)
-            # Add waiting time
-            await asyncio.sleep(1)
+            # Wait till event is set
+            await wait_till_event.wait()
 
     with patch(
         "homeassistant.components.command_line.switch.CommandSwitch",
@@ -676,7 +679,7 @@ async def test_updating_to_often(
                             "command_on": "echo 2",
                             "command_off": "echo 3",
                             "name": "Test",
-                            "scan_interval": 0.1,
+                            "scan_interval": 10,
                         }
                     }
                 ]
@@ -689,12 +692,25 @@ async def test_updating_to_often(
         "Updating Command Line Switch Test took longer than the scheduled update interval"
         not in caplog.text
     )
-    called.clear()
-    caplog.clear()
-
-    async_fire_time_changed(hass, dt_util.now() + timedelta(seconds=1))
+    async_fire_time_changed(hass, dt_util.now() + timedelta(seconds=11))
     await hass.async_block_till_done()
+    assert called
+    called.clear()
 
+    assert (
+        "Updating Command Line Switch Test took longer than the scheduled update interval"
+        not in caplog.text
+    )
+
+    # Simulate update takes too long
+    wait_till_event.clear()
+    async_fire_time_changed(hass, dt_util.now() + timedelta(seconds=10))
+    await asyncio.sleep(0)
+    async_fire_time_changed(hass, dt_util.now() + timedelta(seconds=10))
+    wait_till_event.set()
+
+    # Finish processing update
+    await hass.async_block_till_done()
     assert called
     assert (
         "Updating Command Line Switch Test took longer than the scheduled update interval"